Add artifact handling to BitbucketJenkinsRule
