+++ src/test/java/it/com/atlassian/bitbucket/jenkins/internal/config/BitbucketProjectConfigurationIT.java
@@ -5,12 +5,10 @@ import com.gargoylesoftware.htmlunit.html.*;
 import hudson.model.FreeStyleProject;
 import hudson.plugins.git.BranchSpec;
 import it.com.atlassian.bitbucket.jenkins.internal.fixture.BitbucketJenkinsRule;
-import it.com.atlassian.bitbucket.jenkins.internal.fixture.BitbucketJenkinsWebClientRule;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.RuleChain;
 import org.xml.sax.SAXException;
 
 import java.io.IOException;
@@ -35,16 +33,11 @@ public class BitbucketProjectConfigurationIT {
     @Rule
     public BitbucketJenkinsRule bbJenkinsRule = new BitbucketJenkinsRule();
 
-    @Rule
-    public RuleChain ruleChain = bbJenkinsRule.getRuleChain();
-    private BitbucketJenkinsWebClientRule bitbucketWebClient;
-
     private FreeStyleProject project;
 
     @Before
     public void setup() throws IOException {
         project = bbJenkinsRule.createFreeStyleProject(JENKINS_PROJECT_NAME);
-        bitbucketWebClient = bbJenkinsRule.getWebClientRule();
     }
 
     @After
@@ -56,7 +49,7 @@ public class BitbucketProjectConfigurationIT {
     public void testBitbucketSCMFieldsShouldBePopulatedWithProperValues() throws IOException, SAXException {
         setupBitbucketSCM();
 
-        HtmlPage configurePage = bitbucketWebClient.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
+        HtmlPage configurePage = bbJenkinsRule.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
         HtmlForm form = configurePage.getFormByName("config");
 
         HtmlSelect credential = form.getSelectByName("_.credentialsId");
@@ -73,7 +66,7 @@ public class BitbucketProjectConfigurationIT {
 
     @Test
     public void testCreateBitbucketProject() throws Exception {
-        HtmlPage configurePage = bitbucketWebClient.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
+        HtmlPage configurePage = bbJenkinsRule.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
         HtmlForm form = configurePage.getFormByName("config");
         List<HtmlRadioButtonInput> scms = form.getRadioButtonsByName("scm");
         Optional<HtmlRadioButtonInput> bitbucketSCMRadioButton = scms.stream()
@@ -83,7 +76,7 @@ public class BitbucketProjectConfigurationIT {
         //Configure Bitbucket SCM
         assertTrue(bitbucketSCMRadioButton.isPresent());
         bitbucketSCMRadioButton.get().click();
-        bitbucketWebClient.waitForBackgroundJavaScript();
+        bbJenkinsRule.waitForBackgroundJavaScript();
 
         HtmlSelect credential = form.getSelectByName("_.credentialsId");
         waitTillItemIsRendered(credential::getOptions);
@@ -127,14 +120,14 @@ public class BitbucketProjectConfigurationIT {
     public void testProjectEmpty() throws IOException, SAXException {
         setupBitbucketSCM();
 
-        HtmlPage configurePage = bitbucketWebClient.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
+        HtmlPage configurePage = bbJenkinsRule.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
         HtmlForm form = configurePage.getFormByName("config");
 
         HtmlInput projectNameInput = form.getInputByName("_.projectName");
         projectNameInput.click();
         projectNameInput.setValueAttribute("");
         form.click();
-        bitbucketWebClient.waitForBackgroundJavaScript();
+        bbJenkinsRule.waitForBackgroundJavaScript();
         assertNotNull(getDivByText(form, "Required"));
     }
 
@@ -142,14 +135,14 @@ public class BitbucketProjectConfigurationIT {
     public void testProjectNotExist() throws IOException, SAXException {
         setupBitbucketSCM();
 
-        HtmlPage configurePage = bitbucketWebClient.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
+        HtmlPage configurePage = bbJenkinsRule.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
         HtmlForm form = configurePage.getFormByName("config");
 
         HtmlInput projectNameInput = form.getInputByName("_.projectName");
         projectNameInput.click();
         projectNameInput.setValueAttribute("non-existent-project");
         form.click();
-        bitbucketWebClient.waitForBackgroundJavaScript();
+        bbJenkinsRule.waitForBackgroundJavaScript();
         assertNotNull(getDivByText(form, "The project 'non-existent-project' does not exist or you do not have permission to access it."));
     }
 
@@ -157,7 +150,7 @@ public class BitbucketProjectConfigurationIT {
     public void testRepositoryEmpty() throws Exception {
         setupBitbucketSCM();
 
-        HtmlPage configurePage = bitbucketWebClient.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
+        HtmlPage configurePage = bbJenkinsRule.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
         HtmlForm form = configurePage.getFormByName("config");
         HtmlInput projectNameInput = form.getInputByName("_.projectName");
         projectNameInput.click();
@@ -168,7 +161,7 @@ public class BitbucketProjectConfigurationIT {
         repoNameInput.click();
         repoNameInput.setValueAttribute("");
         form.click();
-        bitbucketWebClient.waitForBackgroundJavaScript();
+        bbJenkinsRule.waitForBackgroundJavaScript();
         assertNotNull(getDivByText(form, "Required"));
     }
 
@@ -176,7 +169,7 @@ public class BitbucketProjectConfigurationIT {
     public void testRepositoryNotExist() throws Exception {
         setupBitbucketSCM();
 
-        HtmlPage configurePage = bitbucketWebClient.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
+        HtmlPage configurePage = bbJenkinsRule.visit("job/" + JENKINS_PROJECT_NAME + "/configure");
         HtmlForm form = configurePage.getFormByName("config");
         HtmlInput projectNameInput = form.getInputByName("_.projectName");
         projectNameInput.click();
@@ -187,7 +180,7 @@ public class BitbucketProjectConfigurationIT {
         repoNameInput.click();
         repoNameInput.setValueAttribute("non-existent-repo");
         form.click();
-        bitbucketWebClient.waitForBackgroundJavaScript();
+        bbJenkinsRule.waitForBackgroundJavaScript();
 
         assertNotNull(getDivByText(form, "The repository 'non-existent-repo' does not exist or you do not have permission to access it."));
     }
